fix cp year, move writer to better package:wq
